// packages/MobileApp/src/navigation/types.ts
export type MonitorsStackParamList = {
  MonitorsList:
    | { initialFilter?: "all" | "issues" | "operational" | "disabled" }
    | undefined;
  MonitorDetail: { monitorId: string; projectId: string };
};